refactor: extract shared upload image helpers into lib/upload-images
HEIC/HEIF normalization and image-file validation were duplicated between
the main upload flow and the board upload page. Move normalizeImageFile,
an isImageFile validator, a formatFileSize helper, and a createIdempotencyKey
helper into apps/web/lib/upload-images.ts and import them in both call sites.
Behavior is unchanged; both create calls keep their existing arguments.
